Move container routes to their own python file
